split repeated_checks.h
