+++ avionics/CMakeLists.txt.in
@@ -1,3 +1,4 @@
+cmake_minimum_required(VERSION 3.10)
 include(ExternalProject)
 ExternalProject_Add(googletest
   GIT_REPOSITORY    https://github.com/google/googletest.git
